Conditionally support vImage based on compile and run time macOS versions
