+++ melatonin/implementations/all.h
@@ -1,15 +1,15 @@
 #pragma once
 
+#include "../../melatonin/cached_blur.h"
 #include "dequeue.h"
-#include "gaussian.h"
 #include "float_vector_stack_blur.h"
+#include "gaussian.h"
 #include "gin.h"
 #include "naive.h"
-#include "naive_with_martin_optimization.h"
 #include "naive_class.h"
+#include "naive_with_martin_optimization.h"
 #include "templated_function.h"
 #include "templated_function_float.h"
 #include "vector.h"
 #include "vector_class.h"
 #include "vector_optimized.h"
-#include "../../melatonin/blur.h"
