+++ tests/drop_shadow.cpp
@@ -332,7 +332,7 @@ TEST_CASE ("convertToARGB static function")
     SECTION ("red at full opacity")
     {
         g2.fillAll (juce::Colours::white); // populate the single channel with all at 255
-        auto converted = melatonin::DropShadow::convertToARGB (singleChannel, juce::Colours::red);
+        auto converted = melatonin::blur::convertToARGB (singleChannel, juce::Colours::red);
         juce::Image::BitmapData data (converted, juce::Image::BitmapData::readWrite);
 
         CHECK (converted.getPixelAt (0, 0).toDisplayString (true) == "FFFF0000");
@@ -347,7 +347,7 @@ TEST_CASE ("convertToARGB static function")
     {
         uint8_t alpha = 85u; // 55 in hex
         g2.fillAll (juce::Colours::white.withAlpha (alpha));
-        auto converted = melatonin::DropShadow::convertToARGB (singleChannel, juce::Colours::red);
+        auto converted = melatonin::blur::convertToARGB (singleChannel, juce::Colours::red);
         juce::Image::BitmapData data (converted, juce::Image::BitmapData::readWrite);
 
         auto actualPixel = getActualARGBPixel (data.getPixelPointer (0, 0));
